Fix filenames that encode to more than 255 bytes in UTF-8
HFS+ names may be up to 255 UTF-16 characters, some of which (e.g. those
in the halfwidth and fullwidth block) encode to 3 bytes in UTF-8,
making the maximum UTF-8 length of any path name 765 bytes.

Until now libhfs effectively only converted the last 255 of these,
resulting in an incorrect name, and hfsfuse itself only supported up to
511 byte names. This fixes the former and corrects the latter to the full
maximum width.

+++ lib/libhfs/unicode.c
@@ -134,7 +134,7 @@ utf16_to_utf8(char *dst, size_t dst_len,
 	      const uint16_t *src, size_t src_len,
 	      int flags, int *errp)
 {
-	uint8_t spos, dpos;
+	size_t spos, dpos;
 	int error;
 
 #define CHECK_LENGTH(l)	(dpos > dst_len-(l) ? dst=NULL : NULL)
